+++ src/dal/migrations/sqlite/Migration20260326211214.ts
@@ -1,9 +1,0 @@
-import { Migration } from '@mikro-orm/migrations';
-
-export class Migration20260326211214 extends Migration {
-
-  override async up(): Promise<void> {
-    this.addSql(`alter table \`outfit\` add column \`slots\` json null;`);
-  }
-
-}